+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+1. `doc.go` package overview documentation.
+2. `set-pc-control` command and emulation.
+
 
 ## [0.8.3](https://github.com/uhppoted/uhppote-simulator/releases/tag/v0.8.3) - 2022-12-16